+++ lib/android_graphics.dart
@@ -4,4 +4,8 @@
 /// and rectangles that let you handle drawing to the screen directly.
 ///
 /// See: https://developer.android.com/reference/android/graphics/package-summary
+///
+/// {@canonicalFor bitmap.Bitmap}
 library android_graphics;
+
+export 'src/graphics/bitmap.dart' show Bitmap;