+++ lib/src/graphics/bitmap.dart
@@ -1,0 +1,9 @@
+/* This is free and unencumbered software released into the public domain. */
+
+import '../os/parcelable.dart' show Parcelable;
+
+/// Bitmap image.
+///
+/// See: https://developer.android.com/reference/android/graphics/Bitmap
+abstract class Bitmap implements Parcelable {
+}